+++ PathObjectivesAndConstraints/src/CBindingHelper.hpp
@@ -9,8 +9,8 @@ class CBindingHelper
 {
     public:
         CBindingHelper();
-        Eigen::Matrix<float,D,4> array_section_to_eigen(float cont_pts[], int &num_cps, unsigned int &index);
-        Eigen::MatrixXf array_to_eigen(float cont_pts[], int &num_cps);
+        Eigen::Matrix<double,D,4> array_section_to_eigen(double cont_pts[], int &num_cps, unsigned int &index);
+        Eigen::MatrixXd array_to_eigen(double cont_pts[], int &num_cps);
 };
 
 #endif
