+++ PathObjectivesAndConstraints/src/ThirdOrderCurvatureBounds.hpp
@@ -11,15 +11,15 @@ class ThirdOrderCurvatureBounds
 {
     public:
         ThirdOrderCurvatureBounds();
-        float get_spline_curvature_bound(float cont_pts[], int &num_control_points);
-        Eigen::VectorXf get_interval_curvature_bounds(float cont_pts[], int &num_control_points);
-        float evaluate_interval_curvature_bound(Eigen::Matrix<float,D,4> &control_points);
+        double get_spline_curvature_bound(double cont_pts[], int &num_control_points);
+        Eigen::VectorXd get_interval_curvature_bounds(double cont_pts[], int &num_control_points);
+        double evaluate_interval_curvature_bound(Eigen::Matrix<double,D,4> &control_points);
     private:
         CBindingHelper<D> cbind_help{};
-        float find_maximum_cross_term(Eigen::Matrix<float,D,4> &control_points, float &scale_factor);
-        Eigen::Vector4d get_2D_cross_coefficients(Eigen::Matrix<float,D,4> &control_points);
-        Eigen::Vector4d get_3D_cross_coefficients(Eigen::Matrix<float,D,4> &control_points);
-        float calculate_cross_term_magnitude(float &t, Eigen::Matrix<float,D,4> &control_points, float &scale_factor);
+        double find_maximum_cross_term(Eigen::Matrix<double,D,4> &control_points, double &scale_factor);
+        Eigen::Vector4d get_2D_cross_coefficients(Eigen::Matrix<double,D,4> &control_points);
+        Eigen::Vector4d get_3D_cross_coefficients(Eigen::Matrix<double,D,4> &control_points);
+        double calculate_cross_term_magnitude(double &t, Eigen::Matrix<double,D,4> &control_points, double &scale_factor);
     FRIEND_TEST(ThirdOrderCurvatureTest, CrossCoeficients2D);
     FRIEND_TEST(ThirdOrderCurvatureTest, CrossCoeficients3D);
     FRIEND_TEST(ThirdOrderCurvatureTest, CrossTermMagnitude);
